requirements.txt
streamlit==1.50.0
pandas==2.3.2
pyarrow==21.0.0
plotly==6.3.0
python-dotenv==1.1.1
